add generic "application/*" mimetype icon
The text-x-script icon that used now is really incorrect for this mimetype,
but there doesn't seem to be an appropriate icon invented for such a
broad range of files (application/* includes file archives as well as
Word documents and other app-specific files).
The right solution would probably be to have a more fine-grained file type
distinction in Claws.